rename fake_client
